+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Bugfixes:
 - should scroll to last row with very high rows respecting fixedRows ([#758](https://github.com/warpech/jquery-handsontable/issues/758))
+- Native Scrollbar was listening to window scroll also after table was removed from DOM
 
 ## [0.9.3](https://github.com/warpech/jquery-handsontable/tree/v0.9.3) (Jun 4, 2013)